drivers: clock_control: Introduce revision 3 of CCM driver
This commit introduces a new version of the CCM driver. The purpose of this
is to make the CCM driver usable by all NXP SoCs despite the changes in the
HAL API.

With this driver, each SoC will have to define its own clock tree and
a few common operations such as clock enable/disable, set_rate and
get_rate.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>